Rename dependency
