+++ src/lib/authik/server.ts
@@ -1,4 +1,4 @@
-import { Authik } from '@authik/next/server'
+import { Authik } from '@authik/nextjs/server'
 
 export const authik = new Authik(
   process.env.AUTHIK_SECRET_KEY,
